Logic for compiling search data when > 1 page of results is returned
Some of GitHub’s endpoints can return more than one page of results.

If more than one page of results are present, GitHub lists that information in the headers.

This class parses that header information to determine if more than one page of results are present.

If they are, this class:

- determines how many pages there are
- repeatedly hits the endpoint, incrementing the page number request, until the last page is reached
- compiles all of the data and returns it in a flattened format